Allow commands to return an exit code
It is important for any program to exit with a non-zero exit code if it
encountered an error.

In Symfony console commands it's common to return with an exit code in
the execute() method of a command. Drupal Console disregards this exit
code in Application::doRun(), making it harder to use Drupal Console
in a scriptable way.