[console] refactor and generalize validating dependencies
Some commands like ModuleCommand and ProfileCommand need to validate
dependencies and are using the same exact code to do it, replicated
per-command; refactor it and share the implementation between the two
commands.

While at it also make the checkDependencies() function more generic, do
not validate "dependencies" (i.e. modules) but rather "extensions",
taking a $type argument to restrict validation to certain extension
types, for instance the new method can be used to validate themes.

The new validating method is Validator::validateExtensions() which
checks machine names and then calls the new method
ExtensionManager::checkExtensions() which checks that the passed machine
names are actually extensions of the given type.

NOTE: since Validator::validateExtension() accepts a string argument add
a default value as an empty string for the '--dependencies' option in
ModuleCommand and ProfileCommand.

The default value is the last argument of addOption() as specified on
http://api.symfony.com/3.1/Symfony/Component/Console/Command/Command.html#method_addOption