[generate:profile] add a '--themes' option (Fixes #3092)
When an installation profile imports configuration which refers to the
installed themes it is necessary that the profile refers to these themes
in the PROFILE.info.yml file too.

Without this the installation can fail with a message like:

  Drupal\Core\Config\UnmetDependenciesException: ...
  PROFILE have unmet dependencies in
  .../web/core/lib/Drupal/Core/Config/UnmetDependenciesException.php:84

So add a '--themes' options to easily specify the enabled themes when
generating a profile.