[generate:profile] add option to specify enabled themes (#3173)
* [console] Validator: rename validateModuleDependencies to validateMachineNameList

The new name describes better what the function does, and it's also more
generic: it can be used to validate any kind of list of extensions
names, not just dependencies (i.e. modules) but also a list of machine
names of themes.

* [console] return $this from ExtensionManager::discoverExtension()

This allows to chain-call show* methods after discoverExtension().

* [console] refactor and generalize validating dependencies

Some commands like ModuleCommand and ProfileCommand need to validate
dependencies and are using the same exact code to do it, replicated
per-command; refactor it and share the implementation between the two
commands.

While at it also make the checkDependencies() function more generic, do
not validate "dependencies" (i.e. modules) but rather "extensions",
taking a $type argument to restrict validation to certain extension
types, for instance the new method can be used to validate themes.

The new validating method is Validator::validateExtensions() which
checks machine names and then calls the new method
ExtensionManager::checkExtensions() which checks that the passed machine
names are actually extensions of the given type.

NOTE: since Validator::validateExtension() accepts a string argument add
a default value as an empty string for the '--dependencies' option in
ModuleCommand and ProfileCommand.

The default value is the last argument of addOption() as specified on
http://api.symfony.com/3.1/Symfony/Component/Console/Command/Command.html#method_addOption

* [console] remove some class members which are now unused

After the checkDependencies() refactoring, ModuleCommand and
ProfileCommand do not need to depend on http_client and console.site
anymore, because the code using them is now in ExtensionManager.

* [console] adjust indentation for some comments

Other instance of wrong indentation could be found with a command like:

  git grep "^\*\/"

* [generate:profile] add a '--themes' option (Fixes #3092)

When an installation profile imports configuration which refers to the
installed themes it is necessary that the profile refers to these themes
in the PROFILE.info.yml file too.

Without this the installation can fail with a message like:

  Drupal\Core\Config\UnmetDependenciesException: ...
  PROFILE have unmet dependencies in
  .../web/core/lib/Drupal/Core/Config/UnmetDependenciesException.php:84

So add a '--themes' options to easily specify the enabled themes when
generating a profile.