[console] Standardize the second argument of Command::addOption(). Fix #3174. (#3286)
The second argument of Symfony\Component\Console\Command\Command::addOption()
is the shortcut name of the option, and it can be set to null when not used,
as reported in the documentation:
http://api.symfony.com/3.1/Symfony/Component/Console/Command/Command.html#method_addOption

However, currently sometimes false or an empty string are used when the option
does not have a shortcut name. This could cause some confusion.

Use null everywhere as the second argument when the option does not have
a shortcut name.